chore: move kaggle pipeline to kaggle package
additionally, a `pipelines kaggle` subcommand was added to the gravy
command.
